requirements.txt
numpy~=1.23.5
pandas~=1.5.3
scipy~=1.10.1
pynwb~=2.2.0
einops~=0.6.0
pytest~=7.2.1
h5py~=3.8.0
six~=1.16.0
matplotlib
sympy~=1.11.1
scikit-learn~=1.2.1
networkx~=3.0
jsonschema~=4.21.1
scikit-image~=0.19.3
tqdm~=4.64.1
setuptools~=60.2.0
bokeh~=3.0.3
PyYAML~=6.0
rich==13.3.2
torch-optimizer==0.3.0
# xformers==0.0.24
absl-py==1.2.0
tensorboard~=2.13
hydra-core~=1.3.2
lightning==2.0.8
wandb~=0.15
tabulate~=0.9
zenodo-get~=1.5.1
pympler~=1.0.1
msgpack~=1.0.5
snakemake~=7.32.3
torchtyping~=0.1
torchtext==0.17.0 # required, from here: https://github.com/pytorch/text#installation
cython
mne~=1.5
pydantic~=2.0
seaborn~=0.13
pulp==2.7.0
black==24.2.0
pre-commit==3.6.2
g2p-en~=2.1.0

# Stella added
future #~=0.15.2
dnspython~=2.0.0
# snakemake==7.32.4 # it will automatically install allensdk
allensdk
cffi
# conda install -c conda-forge dandi


# about torch-things
# pip install xformers==0.0.24 torch==2.2.0 torchvision==0.17.0 torchaudio==2.2.0 --index-url https://download.pytorch.org/whl/cu121
# for snakemake.. mamba is not working :( 

# before you go..
# conda install -c conda-forge gcc
# conda install -c conda-forge c-compiler cxx-compiler